refactor: split local gateway client helpers
